# docs/swagger.yaml
basePath: /v1
definitions:
...
  model.LoginRequest:
    properties:
      password:
        type: string
      user_name:
        type: string
    type: object
  model.ProfileUpdateRequest:
    properties:
      birthday:
        type: string
      first_name:
        type: string
      last_name:
        type: string
      password:
        type: string
    type: object
  model.SignupRequest:
    properties:
      birthday:
        type: string
      email:
        type: string
      first_name:
        type: string
      last_name:
        type: string
      password:
        type: string
      user_name:
        type: string
    type: object
host: localhost:8080
info:
  contact:
    email: support@news-feed.com
    name: API Support
    url: http://news-feed.com/support
  description: This is a sample news feed server.
  license:
    name: MIT
    url: https://opensource.org/licenses/MIT
  termsOfService: http://news-feed.com/terms/
  title: News Feed API
  version: "1.0"
paths:
...
  /v1/users/login:
    post:
      consumes:
      - application/json
      description: Authenticates a user and returns a token.
      parameters:
      - description: User credentials
        in: body
        name: credentials
        required: true
        schema:
          $ref: '#/definitions/model.LoginRequest'
      produces:
      - application/json
      responses:
        "200":
          description: JWT token
          schema:
            additionalProperties:
              type: string
            type: object
        "400":
          description: Invalid request payload
          schema:
            type: string
        "401":
          description: Unauthorized
          schema:
            type: string
      summary: User login
      tags:
      - users
  /v1/users/profile:
    put:
      consumes:
      - application/json
      description: Updates the user profile information.
      parameters:
      - description: Profile update information
        in: body
        name: profileUpdate
        required: true
        schema:
          $ref: '#/definitions/model.ProfileUpdateRequest'
      produces:
      - application/json
      responses:
        "200":
          description: Profile updated successfully
          schema:
            additionalProperties:
              type: string
            type: object
        "400":
          description: Invalid request payload
          schema:
            type: string
        "500":
          description: Internal server error
          schema:
            type: string
      summary: Edit user profile
      tags:
      - users
  /v1/users/signup:
    post:
      consumes:
      - application/json
      description: Registers a new user and returns a token.
      parameters:
      - description: New user signup information
        in: body
        name: signupRequest
        required: true
        schema:
          $ref: '#/definitions/model.SignupRequest'
      produces:
      - application/json
      responses:
        "200":
          description: JWT token
          schema:
            additionalProperties:
              type: string
            type: object
        "400":
          description: Invalid request payload
          schema:
            type: string
        "500":
          description: Internal server error
          schema:
            type: string
      summary: User signup
      tags:
      - users
swagger: "2.0"